// app-server/src/modules/question-paper/update/question-paper-update.dto.ts
export enum QuestionPaperUpdateIdentifierEnum {
  QUESTION_PAPER_META = 'question-paper.meta',
  QUESTION_PAPER_UPDATE = 'question-paper.update',

  // Question
  QUESTION_CREATE = 'question-paper.question.create',
  QUESTION_UPDATE = 'question-paper.question.update',
  QUESTION_DELETE = 'question-paper.question.delete',

  // Option
  OPTION_CREATE = 'question-paper.option.create',
  OPTION_UPDATE = 'question-paper.option.update',
  OPTION_DELETE = 'question-paper.option.delete',

  // Sub Question
  SUB_QUESTION_CREATE = 'question-paper.sub-question.create',
  SUB_QUESTION_UPDATE = 'question-paper.sub-question.update',
  SUB_QUESTION_DELETE = 'question-paper.sub-question.delete',

  // Sub Question Option
  SUB_QUESTION_OPTION_CREATE = 'question-paper.sub-question.option.create',
  SUB_QUESTION_OPTION_UPDATE = 'question-paper.sub-question.option.update',
  SUB_QUESTION_OPTION_DELETE = 'question-paper.sub-question.option.delete',
}